\name{apoptosisQuantification-package}
\alias{apoptosisQuantification-package}
\alias{apoptosisQuantification}
\docType{package}
\title{
\packageTitle{apoptosisQuantification}
}
\description{
\packageDescription{apoptosisQuantification}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{apoptosisQuantification}
\packageIndices{apoptosisQuantification}
}
\author{
\packageAuthor{apoptosisQuantification}

Maintainer: \packageMaintainer{apoptosisQuantification}
}
\references{
Tanzer et al. (2020): Quantitative and Dynamic Catalogs of Proteins 
Released during Apoptotic and Necroptotic Cell Death. 
\emph{Cell Reports}, 30,  1260-1270.e5. 10.1016/j.celrep.2019.12.079.
}
