% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaProportion.R
\name{calculateProportionOfMitochondrialProteins}
\alias{calculateProportionOfMitochondrialProteins}
\title{Calculate contribution of mitochondrial proteins}
\usage{
calculateProportionOfMitochondrialProteins(prot, id = c("Symbol", "Ensembl"))
}
\arguments{
\item{prot}{matrix, data.frame, or tibble}

\item{id}{character(1), either \code{"Symbol"} or \code{"Ensembl"}}
}
\value{
numeric vector
}
\description{
Apoptotic cells show an increase in the proportion of gene expression
which originates from the mitochondria, which occurs after the opening of 
the MPT pore as apoptosis begins. This gradually results in a decrease in 
the total expression of chromosomal genes as the DNA is degraded. 
In late apoptosis, there is typically a drastically reduced total count of 
genes in a cell, with a large proportion of the gene expression stemming 
from mitochondria. 

For transcriptomics, typically, at a threshold of where 10-20% of 
transcripts are mitochondrial in origin, the total number of transcripts 
drops substantially (<20% of what it was prior to apoptosis).

The function \code{calculateProportionOfMitochondrialProteins} calculates
proportions of the intensities of mitochondrial proteins (i.e. proteins that 
are encoded in the mitochondrial DNA) to all intensities. The function 
returns the proportions in percent. 

Mitochondrial proteins are defined by the Human Protein Atlas 
(1139 proteins found by searching the Human Protein Atlas database for 
mitochondrial sublocalization). These proteins are matched against the
proteins found in \code{prot}. Currently, two types of ids are 
encoded: either \code{rownames(prot)} have to be "Symbol" or "Ensembl" ids.
}
\details{
Protein intensities have to in untransformed format such that 
proportions can be calculated as are.

Protein ids are stored in \code{rownames(prot)}. In case of multiple
assignments the ids are separeted by \code{";"}. In that case, a 
feature will be treated as of mitochondrial origin when there is at least
one mitochondrial protein in the multiple assignment.
}
\examples{
set.seed(2022) 
## fake measured protein intensities and create matrix that stores 
## intensities
prot_vals <- runif(1000, min = 10000, max = 1000000)
prot <- matrix(prot_vals, ncol = 10, nrow = 100)

## fake some protein names and sample names
rownames(prot) <- paste("prot", seq_len(nrow(prot)), sep = "_")
colnames(prot) <- paste("sample", seq_len(ncol(prot)), sep = "_")

## randomly assign 10 mitochondrial proteins to the matrix
inds_mito <- sample(seq_len(nrow(prot)), 10)
prot_mito <- c("MDH2", "OMA1", "NLN", "CS", "OPA1", "CSPG5", "SPATA9", 
    "POLG", "RAB38", "NSD3")
rownames(prot)[inds_mito] <- prot_mito

calculateProportionOfMitochondrialProteins(prot = prot, id = "Symbol")
}
\references{
The Human Protein Atlas 
(search: subcell_location:Mitochondria, 
https://www.proteinatlas.org/humanproteome/subcellular/mitochondria, 
accessed January 31, 2022).

Uhlen et al. (2015): Tissue-based map of the human proteome.
\emph{Science}, 347, 1260419. 10.1126/science.1260419

Thul et al. (2017): A subcellular map of the human proteome.
\emph{Science}, 356, eaal3321. 10.1126/science.aal3321
}
