% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaLoadings.R
\name{combineLoadings}
\alias{combineLoadings}
\title{Combine the loadings/calculate the length of loading vectors}
\usage{
combineLoadings(loadings, PC = c("PC1", "PC2"))
}
\arguments{
\item{loadings}{matrix or data.frame}

\item{PC}{character}
}
\value{
numeric vector
}
\description{
The function \code{combineLoadings} calculates the length of loading vectors
across several principal components within the Cartesian coordinate system. 

The length of vectors is calculated along the principal coordinates
defined by \code{PC}, e.g. if \code{PC = c("PC1", "PC2")} the 
length is calculated along the loadings of principal components 1 and 2.
}
\details{
The length of vectors is calculated using the classical way in the 
cartesian coordinate system. 

For 1D (e.g. length of vectors for PC1):
\deqn{
    \bar{\bar{a}} = \sqrt{a_1^2}
}

For 2D (e.g. length of vectors for PC1 and PC2): 
\deqn{
    \bar{\bar{a}} = \sqrt{a_1^2 + a_2^2}
}

For 3D (e.g. length of vectors for PC1, PC2, and PC3):
\deqn{
    \bar{\bar{a}} = \sqrt{a_1^2 + a_2^2 + a_3^2}
}

where $a_1$ is the loading of a data point for PC1, $a_2$ for PC2, $a_3$ for 
PC3, etc.
}
\examples{
library(MatrixQCvis)
library(SummarizedExperiment)
library(tibble)

f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
se <- readRDS(f)
prot <- assay(se)
loadings <- prcomp(t(prot))$rotation |> 
    as.data.frame() |>
    rownames_to_column(var = "protein") |> 
    as_tibble()
combineLoadings(loadings = loadings, PC = c("PC1", "PC2")) 
}
