% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaLoadings.R
\name{createLoadingsTbl}
\alias{createLoadingsTbl}
\title{Create a tibble containing the loading values for marker and 
non-marker features}
\usage{
createLoadingsTbl(vals_markers = vals_markers, vals_all = vals_all)
}
\arguments{
\item{vals_markers}{\code{numeric} vector}

\item{vals_all}{\code{numeric} vector}
}
\value{
tibble
}
\description{
The function \code{createLoadingsTbl} creates a tibble containing the loading
values for marker and non-marker features. 

\code{createLoadingsTbl} returns a tibble with three columns:
\itemize{
    \item \code{feature}: name of feature (e.g. protein, in case of multiple
         assignment separated by \code{";"}),
    \item \code{type}: either \code{"marker"} or \code{non-marker}
}
}
\details{
In case of multiple assignments of features (e.g. proteins), such a multiple 
assignment is treated as a marker feature if there is at least one marker 
feature (defined in \code{names(vals_markers)}) in the multiple assignment.
}
\examples{
library(MatrixQCvis)
library(SummarizedExperiment)

## load data set of Tanzer et al. (2020)
## source: https://www.ebi.ac.uk/pride/archive/projects/PXD014966
f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
tanzer2020 <- readRDS(f)

## get loadings of markers and of the complete data set
prot <- assay(tanzer2020)
markers <- readMarkers(type = "apoptosis")
loadings_markers <- apoptosisQuantification:::getLoadingsOfMarkers(
    values = prot, markers = markers)
vals_all <- prcomp(t(prot))$rotation[, 1]
vals_markers <- setNames(loadings_markers$PC1, loadings_markers$feature)
apoptosisQuantification:::createLoadingsTbl(vals_markers = vals_markers, 
    vals_all = vals_all)
}
