% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaDecoupleR.R
\name{decoupleProteins}
\alias{decoupleProteins}
\title{Calculate score values using decoupleR/run_wmean}
\usage{
decoupleProteins(
  values,
  args_fct,
  contamination = c("apoptosis", "necroptosis")
)
}
\arguments{
\item{values}{matrix, the rownames contain the protein names (SYMBOL ids) and
colnames the samples}

\item{args_fct}{list, containing the entries \code{"n_perm"} 
(\code{numeric(1)} and \code{"signatures"} (\code{tibble})}

\item{contamination}{\code{character(1)}, either \code{"apoptosis"} or
\code{"necroptosis"}}
}
\value{
tibble
}
\description{
The function \code{decoupleProteins} calculates contamination scores
using the \code{run_wmean} function from the \code{decoupleR}. The 
function will return score values (per sample) that can be interpreted as the 
standard deviations away from an empirical null distribution.
}
\details{
The function is insprired by the work of Aurelien Dugourd and his 
plasmaContamination package 
(https://github.com/saezlab/plasmaContamination/).

The function will be called in \code{scoreSamples}.
}
\examples{
library(SummarizedExperiment)
f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
tanzer2020 <- readRDS(f)
prot <- assay(tanzer2020) |>
    as.data.frame()

## define a list with paramters
args_fct <- list()
args_fct[["signatures"]] <- readMarkers(type = "apoptosis", fc = 2, n = 1)
args_fct[["n_perm"]] <- 100

## run the function
decoupleProteins(values = prot, args_fct = args_fct, 
    contamination = "apoptosis")
}
