% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaLoadings.R
\name{getLoadingsOfMarkers}
\alias{getLoadingsOfMarkers}
\title{Get PCA loadings of marker proteins}
\usage{
getLoadingsOfMarkers(prot, markers)
}
\arguments{
\item{prot}{matrix, data.frame, or tibble}

\item{markers}{tibble}
}
\value{
tibble
}
\description{
The function \code{getLoadingsOfMarkers} retrieves the loadings of the
marker proteins (defined by \code{markers}) found in \code{prot}.

\code{getLoadingsOfMarkers} returns a \code{tibble} of the marker 
proteins found in \code{prot} containing the loadings for the principal
components.
}
\details{
In case there are ambiguous/multiple assignments of proteins (containing 
markers) to a loading value (MaxQuant separates these assignments by ";"), 
the proteins that are not in \code{markers$protein} are removed.
In case there are multiple assignments left of marker proteins, the
mean value of the loading values are returned.
}
\examples{
library(MatrixQCvis)
library(SummarizedExperiment)

## "apoptosis"
markers <- readMarkers(type = "apoptosis")
f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
se <- readRDS(f)
prot <- assay(se) |>
    imputeAssay(method = "MinDet")
getLoadingsOfMarkers(prot = prot, markers = markers)

## "nectroptosis"
markers <- readMarkers(type = "necroptosis")
getLoadingsOfMarkers(prot = prot, markers = markers)
}
\references{
Tanzer et al. (2020): Quantitative and Dynamic Catalogs of Proteins 
Released during Apoptotic and Necroptotic Cell Death. 
\emph{Cell Reports}, 30, 1260-1270.e5. 10.1016/j.celrep.2019.12.079.
}
