% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaLoadings.R
\name{performWilcoxonTest}
\alias{performWilcoxonTest}
\title{Perform the Wilcoxon rank sum and signed rank tests on loading values}
\usage{
performWilcoxonTest(tbl)
}
\arguments{
\item{tbl}{tibble}
}
\value{
htest
}
\description{
The function \code{performWilcoxonTest} performs a Wilcoxon rank sum and 
signed rank test on the loading values. The null hypothesis is that 
the mean of ranks of the marker loadings is smaller or equal to the 
ranks of the non-marker loadings. 
The null hypothesis can be rejected with probability \code{p}
(alternative hypothesis: loadings of markers are greater than the loadings
of non-markers).
}
\details{
The function \code{performWilcoxonTest} takes as input a \code{tibble} 
(\code{tbl}) that stores the loading values in the column \code{"value"} 
and the information on marker/non-marker in the column \code{"type"}.
}
\examples{
library(MatrixQCvis)
library(SummarizedExperiment)

## load data set of Tanzer et al. (2020)
## source: https://www.ebi.ac.uk/pride/archive/projects/PXD014966
f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
tanzer2020 <- readRDS(f)

## get loadings of markers and of the complete data set
prot <- assay(tanzer2020)
markers <- readMarkers(type = "apoptosis")
loadings_markers <- getLoadingsOfMarkers(prot = prot, markers = markers)
vals_all <- prcomp(t(prot))$rotation[, 1]
vals_markers <- setNames(loadings_markers$PC1, loadings_markers$protein)
tbl <- createLoadingsTbl(vals_markers = vals_markers, vals_all = vals_all)

performWilcoxonTest(tbl)
}
