% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaDecoupleR.R
\name{permuteFeatures}
\alias{permuteFeatures}
\title{Calculate score values by calculating difference between
means of permuted features and markers}
\usage{
permuteFeatures(
  values,
  args_fct,
  contamination = c("apoptosis", "necroptosis"),
  seed = 2022
)
}
\arguments{
\item{values}{\code{data.frame}}

\item{args_fct}{list of arguments, has to contain the entries
\code{n_perm} (\code{numeric(1)}) and \code{signatures} (\code{tibble})}

\item{contamination}{\code{character(1)}, either \code{"apoptosis"} or 
\code{"necroptosis"}}

\item{seed}{\code{numeric(1)}}
}
\description{
The function \code{permuteFeatures} calculates the difference between the
mean of marker features (e.g. proteins) and random samples of the 
feature universe (features in \code{values}). The scores will be a 
relative quantification of the marker features to the universe 
feature intensities. Samples that 
undergo apoptosis (or any other contamination) will have different 
differences than uncontaminated samples.

The output will be a \code{tibble} containing the differences in means
for the samples and repetitions.
}
\details{
The number of repetitions is defined by \code{n_perm}.

The function will be called in \code{scoreSamples}.
}
\examples{
library(SummarizedExperiment)
f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
tanzer2020 <- readRDS(f)
prot <- assay(tanzer2020) |>
    as.data.frame()

## define a list with paramters
args_fct <- list()
args_fct[["signatures"]] <- readMarkers(type = "apoptosis", fc = 2, n = 1)
args_fct[["n_perm"]] <- 100

## run the function
permuteFeatures(values = prot, args_fct = args_fct, 
    contamination = "apoptosis", seed = 2022)
}
