% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaDecoupleR.R
\name{permuteProteins}
\alias{permuteProteins}
\title{Calculate score values by calculating difference between
means of permuted proteins and markers}
\usage{
permuteProteins(
  values,
  args_fct,
  contamination = c("apoptosis", "necroptosis"),
  seed = 2022
)
}
\arguments{
\item{values}{\code{data.frame}}

\item{args_fct}{list of arguments, has to contain the entries
\code{n_perm} (\code{numeric(1)}) and \code{signatures} (\code{tibble})}

\item{contamination}{\code{character(1)}, either \code{"apoptosis"} or 
\code{"necroptosis"}}

\item{seed}{\code{numeric(1)}}
}
\description{
The function \code{permuteProteins} calculates the difference
}
\details{
The function will be called in \code{scoreSamples}.
}
\examples{
library(SummarizedExperiment)
f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
tanzer2020 <- readRDS(f)
prot <- assay(tanzer2020) |>
    as.data.frame()

## define a list with paramters
args_fct <- list()
args_fct[["signatures"]] <- readMarkers(type = "apoptosis", fc = 2, n = 1)
args_fct[["n_perm"]] <- 100

## run the function
permuteProteins(values = prot, args_fct = args_fct, 
    contamination = "apoptosis", seed = 2022)
}
