% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaLoadings.R
\name{plotPCAandLoadings}
\alias{plotPCAandLoadings}
\title{Plot PCA and loadings of markers and non-markers}
\usage{
plotPCAandLoadings(values, markers, x_coord = "PC1", y_coord = "PC2")
}
\arguments{
\item{values}{\code{matrix} or \code{data.frame}}

\item{markers}{\code{tibble}}

\item{x_coord}{\code{character(1)}, specifying the principal component on the
x axis}

\item{y_coord}{\code{character(1)}, specifying the principal component on the
y axis}
}
\value{
\code{plotly}
}
\description{
The function creates a PCA plot and a plot of the loadings. The markers
and non-markers are colour-coded in the loadings plot.
}
\details{
The function \code{plotPCAandLoadings} takes as input a \code{matrix}, 
\code{data.frame}, or \code{tibble} (\code{values}) 
that stores the intensities of samples column \code{"value"} and the 
information on marker/non-marker in the column \code{"type"}.
}
\examples{
library(MatrixQCvis)
library(SummarizedExperiment)

## load data set of Tanzer et al. (2020)
## source: https://www.ebi.ac.uk/pride/archive/projects/PXD014966
f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
tanzer2020 <- readRDS(f)

## get the intensities and markers
values <- assay(tanzer2020)
markers <- readMarkers(type = "apoptosis")

## plot the PCA and loadings plot
plotPCAandLoadings(values = values, markers = markers)
}
