% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaProportion.R
\name{plotProportionOfMitochondrialProteins}
\alias{plotProportionOfMitochondrialProteins}
\title{Plot proportion of mitochondrial proteins}
\usage{
plotProportionOfMitochondrialProteins(proportion)
}
\arguments{
\item{proportion}{numeric vector}
}
\value{
\code{gg}
}
\description{
The function \code{plotProportionOfMitochondrialProteins} creates a 
plot of the proportion of mitochondrial proteins 
(obtained via \code{calculateProportionOfMitochondrialProtein}).
}
\details{
\code{names(proportion)} are taken as x-axis labels. 

If there are no names for \code{proportion} 
(\code{names(proportion)} is NULL), names from \code{1:length(proportion)}
will be assigned to \code{proportion}.
}
\examples{
set.seed(1)
## fake measured protein intensities and create matrix that stores 
## intensities
prot_vals <- runif(1000, min = 10000, max = 1000000)
prot <- matrix(prot_vals, ncol = 10, nrow = 100)

## fake some protein names and sample names
rownames(prot) <- paste("prot", seq_len(nrow(prot)), sep = "_")
colnames(prot) <- paste("sample", seq_len(ncol(prot)), sep = "_") 

## randomly assign 10 mitochondrial proteins to the matrix
inds_mito <- sample(seq_len(nrow(prot)), 10)
prot_mito <- c("MDH2", "OMA1", "NLN", "CS", "OPA1", "CSPG5", "SPATA9", "POLG", "RAB38", "NSD3")
rownames(prot)[inds_mito] <- prot_mito

prop_mito <- calculateProportionOfMitochondrialProteins(prot = prot, id = "Symbol")
plotProportionOfMitochondrialProteins(proportion = prop_mito)
}
