% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaLoadings.R
\name{plotViolin}
\alias{plotViolin}
\title{Plot the distribution of loading values via a violin plot}
\usage{
plotViolin(values, markers, PC = "PC1")
}
\arguments{
\item{values}{\code{matrix} or \code{data.frame}}

\item{markers}{tbl}

\item{PC}{character, has to be in the format \code{"PC1", "PC2", "PC3", ...},
the specification of multiple principal components is possible and the
length of loadings vectors will be calculated using these principal 
components}
}
\value{
\code{gg}
}
\description{
The function \code{plotViolin} creates violin plots of the loadings
distribution of marker and non-markers.
}
\details{
The function \code{plotViolin} takes as input a \code{tibble} (\code{tbl}) 
that stores the loading values in the column \code{"value"} and the 
information on marker/non-marker in the column \code{"type"}.
}
\examples{
library(MatrixQCvis)
library(SummarizedExperiment)

## load data set of Tanzer et al. (2020)
## source: https://www.ebi.ac.uk/pride/archive/projects/PXD014966
f <- system.file("protein_datasets/tanzer2020.RDS",
    package = "apoptosisQuantification")
tanzer2020 <- readRDS(f)

## get loadings of markers and of the complete data set
values <- assay(tanzer2020)
markers <- readMarkers(type = "apoptosis", fc = 2, n = 1)

plotViolin(values = values, markers = markers, PC = "PC1")
}
