% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaLoadings.R
\name{prepareTbl}
\alias{prepareTbl}
\title{Helper function to create the output of createLoadingsTbl}
\usage{
prepareTbl(values, markers, PC = "PC1")
}
\arguments{
\item{values}{\code{matrix} or \code{data.frame}}

\item{markers}{tbl}

\item{PC}{character, has to be in the format \code{"PC1", "PC2", "PC3", ...},
the specification of multiple principal components is possible and the
length of loadings vectors will be calculated using these principal 
components}
}
\description{
The function \code{prepareTbl} prepares the input for 
\code{createLoadingsTbl} using the intensity values (\code{values}) and
the \code{tibble} with markers (\code{markers}).
}
\details{
The function will call internally the helper function 
\code{combineLoadings}. This means the function will return absolute 
loadings values even in the case of only one selected principal component.
}
\examples{
library(MatrixQCvis)
library(SummarizedExperiment)

## load data set of Tanzer et al. (2020)
## source: https://www.ebi.ac.uk/pride/archive/projects/PXD014966
f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
tanzer2020 <- readRDS(f)

## get loadings of markers and of the complete data set
prot <- assay(tanzer2020)
markers <- readMarkers(type = "apoptosis")

apoptosisQuantification:::prepareTbl(values = prot, markers = markers, 
    PC = "PC1")
}
