% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaDecoupleR.R
\name{scoreSamples}
\alias{scoreSamples}
\title{scoreSamples}
\usage{
scoreSamples(values, contamination = c("apoptosis", "necroptosis"), ...)
}
\arguments{
\item{values}{numeric \code{vector}, \code{matrix}, or \code{data.frame}}

\item{contamination}{\code{character(1)}, either \code{"apoptosis"} or 
\code{"necroptosis"}}

\item{...}{further arguments passed to \code{readMarkers}}
}
\value{
data.frame, containining normalized enrichment (mean) scores for the
contamination signature (the higher the score, the higher the potential
contamination). If there is no quantitative information on changes 
(e.g. fold changes or t-values), the function will return an empirical
distribution of difference of means of the markers to all features in the 
data set
}
\description{
This function performs an enrichment analysis to test whether a given
proteomic sample is contaminated with markers from apoptosis or necroptosis.
}
\details{
The parameter \code{values} has to have (gene) symbols as \code{row.names}.
\code{values} can be either a numeric \code{vector}, \code{matrix}, or 
\code{data.frame}. It holds the normalized intensity values of features 
(e.g. proteins).

The signatures are taken per default from Tanzer et al. (2020), but also 
a \code{tibble} containing the contamination markers can be given to the 
function as argument \code{signatures}. 
In that case the \code{tibble} has to contain at least the 
column \code{"feature"} that contains the marker features (e.g. proteins).  

The function will either run \code{decoupleFeatures} or 
\code{permuteFeatures} to calculate scores. The behaviour will be triggered
by the presence of the column \code{value} in \code{signatures} (in that 
case, \code{decoupleFeatures} is run).
}
\examples{
library(SummarizedExperiment)

f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
tanzer2020 <- readRDS(f)
prot <- assay(tanzer2020)

## use the intensities to calculate scores, the function will load the 
## data set of Tanzer et al. (2020)
scoreSamples(values = prot, contamination = "apoptosis", n_perm = 100)

## use the contamination scores of Tanzer et al. (2020), but with specified
## fc and n_significan
}
\references{
Tanzer et al. (2020): Quantitative and Dynamic Catalogs of Proteins 
Released during Apoptotic and Necroptotic Cell Death. 
\emph{Cell Reports}, 30,  1260-1270.e5. 10.1016/j.celrep.2019.12.079.
}
