% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaLoadings.R
\name{splitNames}
\alias{splitNames}
\title{Split the names when separated by ;}
\usage{
splitNames(feature_names, na.rm = TRUE)
}
\arguments{
\item{feature_names}{character}

\item{na.rm}{logical(1)}
}
\description{
The function \code{splitNames} creates a ";"-separated list of feature names. 
If \code{na.rm = TRUE}, the \code{NA} values within the feature names 
are removed. \code{NA} values might occur if feature names are translated
from one id to the other, e.g. from SYMBOL to Ensembl.
}
\details{
Helper function for various functions in the \code{apoptosisQuantification}
package.
}
\examples{
protein_names <- c("prot_1;prot_2;NA")
apoptosisQuantification:::splitNames(feature_names = protein_names,
    na.rm = FALSE)
apoptosisQuantification:::splitNames(feature_names = protein_names,
    na.rm = TRUE)
}
