% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaLoadings.R
\name{getLoadingsOfMarkers}
\alias{getLoadingsOfMarkers}
\title{Get PCA loadings of marker features}
\usage{
getLoadingsOfMarkers(values, markers)
}
\arguments{
\item{values}{\code{matrix} or \code{data.frame}}

\item{markers}{tbl}
}
\value{
tibble
}
\description{
The function \code{getLoadingsOfMarkers} retrieves the loadings of the
marker features (defined by \code{markers}) of the \code{values} data set.

\code{getLoadingsOfMarkers} returns a \code{tibble} of the marker 
features found in \code{values} containing the loadings for the principal
components.
}
\details{
In case there are ambiguous/multiple assignments of features (containing 
markers) to a loading value (MaxQuant separates these assignments by ";"), 
the features that are not in \code{markers$feature} are removed.
In case there are multiple assignments left of marker features, the
mean value of the loading values are returned.
}
\examples{
library(MatrixQCvis)
library(SummarizedExperiment)

## "apoptosis"
markers <- readMarkers(type = "apoptosis")
f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
se <- readRDS(f)
prot <- assay(se) |>
    imputeAssay(method = "MinDet")
apoptosisQuantification:::getLoadingsOfMarkers(values = prot, 
    markers = markers)

## "nectroptosis"
markers <- readMarkers(type = "necroptosis")
apoptosisQuantification:::getLoadingsOfMarkers(values = prot, 
    markers = markers)
}
\references{
Tanzer et al. (2020): Quantitative and Dynamic Catalogs of Proteins 
Released during Apoptotic and Necroptotic Cell Death. 
\emph{Cell Reports}, 30, 1260-1270.e5. 10.1016/j.celrep.2019.12.079.
}
