% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaLoadings.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Plot the histogram (distribution) of loading values}
\usage{
plotHistogram(tbl)
}
\arguments{
\item{tbl}{tibble}
}
\value{
\code{gg}
}
\description{
The function \code{plotHistogram} creates (overlaid) histograms
plots of the loadings distribution of marker and non-markers.
}
\details{
The function \code{plotHistogram} takes as input a \code{tibble} (\code{tbl}) 
that stores the loading values in the column \code{"value"} and the 
information on marker/non-marker in the column \code{"type"}.
}
\examples{
library(MatrixQCvis)
library(SummarizedExperiment)

## load data set of Tanzer et al. (2020)
## source: https://www.ebi.ac.uk/pride/archive/projects/PXD014966
f <- system.file("protein_datasets/tanzer2020.RDS",
    package = "apoptosisQuantification")
tanzer2020 <- readRDS(f)

## get loadings of markers and of the complete data set
prot <- assay(tanzer2020)
markers <- readMarkers(type = "apoptosis", fc = 2, n = 1)
loadings_markers <- getLoadingsOfMarkers(prot = prot, markers = markers)
vals_all <- prcomp(t(prot))$rotation[, 2]
vals_markers <- setNames(loadings_markers$PC2, loadings_markers$protein)
tbl <- createLoadingsTbl(vals_markers = vals_markers, vals_all = vals_all)

plotHistogram(tbl)
}
