% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{readMarkers}
\alias{readMarkers}
\title{Load a file with known markers linked to apoptosis or necroptosis}
\usage{
readMarkers(type = c("apoptosis", "necroptosis"), fc = 2, n = 1)
}
\arguments{
\item{type}{character, \code{"apoptosis"} or \code{"necroptosis"}}

\item{fc}{numeric(1), threshold for fold change (default 2), e.g. if set to
2, the proteins are retained that have absolute fold changes greater or equal
to \code{fc}}

\item{n}{numeric, threshold for number of significant time points
(default 1), e.g. if set to 1, the proteins are retained that have at 
least 1 significant time point}
}
\value{
tibble
}
\description{
The function \code{readMarkers} loads a data set with markers linked to
apoptosis or necroptosis. 

The returned \code{tibble} contains three columns:
\itemize{
    \item \code{protein}: name of the marker, 
    \item \code{change}: maximum/minimum fold change from the 
        time-course experiment (maximum absolute value is taken), 
    \item \code{significant_n}: number of significant time-points.
}

Classical necrotic markers include lactate deydrogenase (LDH),
high-mobility group B1 (HMGB1), myoglobin, enolase, and 14-3-3 proteins.
Commonly, a protein release is observed during apoptosis and necroptosis
for human myeloid/lymphoma cell lines and human primary macrophage cells as
observed time course experiments (cf. Tanzer et al., 2020).
}
\details{
The function loads the Supplementary Table S1 from Tanzer et al. (2020).

In the data set of Tanzer et al. (2020), the following abbreviations are 
used:
\itemize{
    \item TNF: tumor necrosis factor (TNF-mediated apoptosis),
    \item SM: birinapant  (inhibition of cIAPs using small molecules called 
        Smac mimetics),
    \item IDN: caspase inhibitor IDUN-6556 (IDN-6556, used in induction of 
        necroptosis),
    \item TNF application: leads to production of cytokines (no induction of 
        apoptosis or necroptosis),
    \item TNF+SM application: inducton of apptosis,
    \item TNF+SM+IDN-6556 application: induction of necroptosis.
}

The time points 1 h, 3 h, 5 h, and 7 h are taken into account for defining
the markers. For apoptosis, the time points 9 h, 12.5 h and 15 h are 
removed prior to filtering the data set.
}
\examples{
## "apoptosis"
readMarkers(type = "apoptosis", fc = 2, n = 1)

## "nectroptosis"
readMarkers(type = "necroptosis", fc = 2, n = 2)
}
\references{
Tanzer et al. (2020): Quantitative and Dynamic Catalogs of Proteins 
Released during Apoptotic and Necroptotic Cell Death. 
\emph{Cell Reports}, 30,  1260-1270.e5. 10.1016/j.celrep.2019.12.079.
}
