% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyViaDecoupleR.R
\name{scoreSamples}
\alias{scoreSamples}
\title{scoreSamples}
\usage{
scoreSamples(values, contamination = c("apoptosis", "necroptosis"), ...)
}
\arguments{
\item{values}{numeric \code{vector}, \code{matrix}, or \code{data.frame}}

\item{contamination}{\code{character(1)}, either \code{"apoptosis"} or 
\code{"necroptosis"}}

\item{...}{further arguments passed to \code{readMarkers}}
}
\value{
data.frame, containining normalized enrichment (mean) scores for the
contamination signature (the higher the score, the higher the potential
contamination)
}
\description{
This function performs an enrichment analysis to test whether a given
proteomic sample is contaminated with markers from apoptosis or necroptosis.
The signatures are taken from Tanzer et al. (2020).
}
\details{
The function is insprired by the work of Aurelien Dugourd and his 
plasmaContamination package 
(https://github.com/saezlab/plasmaContamination/).

The parameter \code{values} has to have (gene) symbols as \code{row.names}.
\code{values} can be either a numeric \code{vector}, \code{matrix}, or 
\code{data.frame}. It can hold the normalized
}
\examples{
library(SummarizedExperiment)

f <- system.file("protein_datasets/tanzer2020.RDS", 
    package = "apoptosisQuantification")
tanzer2020 <- readRDS(f)
prot <- assay(tanzer2020)

## use the intensities to calculate scores
scoreSamples(values = prot, contamination = "apoptosis", n_perm = 100)
}
\references{
Tanzer et al. (2020): Quantitative and Dynamic Catalogs of Proteins 
Released during Apoptotic and Necroptotic Cell Death. 
\emph{Cell Reports}, 30,  1260-1270.e5. 10.1016/j.celrep.2019.12.079.
}
